requirements.txt
Flask[async]
requests
aiohttp
googleapis-common-protos
pycryptodome
protobuf
Werkzeug
pytz
PyGithub
